# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_opentrusty_opentrusty_internal_oauth2.Error:
    properties:
      error:
        type: string
      error_description:
        type: string
      error_uri:
        type: string
      state:
        type: string
    type: object
  github_com_opentrusty_opentrusty_internal_oauth2.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      id_token:
        description: Added for OIDC (Phase II.1)
        type: string
      refresh_token:
        type: string
      scope:
        type: string
      token_type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Production-grade Identity Provider
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: OpenTrusty API
  version: 1.0.0
paths:
...
  /oauth2/authorize:
    get:
      consumes:
      - application/json
      description: Starts the authorization flow (RFC 6749)
      parameters:
      - description: Client ID
        in: query
        name: client_id
        required: true
        type: string
      - description: Redirect URI
        in: query
        name: redirect_uri
        required: true
        type: string
      - description: Response Type (must be 'code')
        in: query
        name: response_type
        required: true
        type: string
      - description: Scopes
        in: query
        name: scope
        type: string
      - description: Random State
        in: query
        name: state
        required: true
        type: string
      - description: Nonce (OIDC)
        in: query
        name: nonce
        type: string
      - description: PKCE Challenge
        in: query
        name: code_challenge
        type: string
      - description: PKCE Method (S256)
        in: query
        name: code_challenge_method
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: Redirects to callback or login
          schema:
            type: string
      summary: OAuth2 Authorize Endpoint
      tags:
      - OAuth2
  /oauth2/revoke:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Revoke a refresh token (RFC 7009)
      parameters:
      - description: Token to revoke
        in: formData
        name: token
        required: true
        type: string
      - description: Client ID
        in: formData
        name: client_id
        type: string
      - description: Client Secret
        in: formData
        name: client_secret
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_opentrusty_opentrusty_internal_oauth2.Error'
      summary: Revoke Token
      tags:
      - OAuth2
  /oauth2/token:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Exchange code for access token (RFC 6749)
      parameters:
      - description: Grant Type (authorization_code or refresh_token)
        in: formData
        name: grant_type
        required: true
        type: string
      - description: Authorization Code (for authorization_code grant)
        in: formData
        name: code
        type: string
      - description: Redirect URI
        in: formData
        name: redirect_uri
        type: string
      - description: Client ID (if not Basic Auth)
        in: formData
        name: client_id
        type: string
      - description: Client Secret (if not Basic Auth)
        in: formData
        name: client_secret
        type: string
      - description: PKCE Verifier
        in: formData
        name: code_verifier
        type: string
      - description: Refresh Token (for refresh_token grant)
        in: formData
        name: refresh_token
        type: string
      - description: Scope
        in: formData
        name: scope
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_opentrusty_opentrusty_internal_oauth2.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_opentrusty_opentrusty_internal_oauth2.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_opentrusty_opentrusty_internal_oauth2.Error'
      summary: OAuth2 Token Endpoint
      tags:
      - OAuth2
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"